+++ src/Settings/SettingsSinkExtensions.cs
@@ -12,21 +12,21 @@ namespace Phoenix.Functionality.Settings;
 public static class SettingsSinkExtensions
 {
 	/// <summary>
-	/// Gets the name for a settings class.
+	/// Gets the name of a settings class respecting the <see cref="SettingsNameAttribute"/>.
 	/// </summary>
 	/// <typeparam name="TSettings"> The generic type of the settings class. </typeparam>
 	/// <param name="sink"> The <see cref="ISettingsSink"/> that is extended. </param>
 	/// <returns> The name of the settings class. </returns>
-	public static string GetSettingsFileNameWithoutExtension<TSettings>(this ISettingsSink sink) where TSettings : ISettings
-		=> sink.GetSettingsFileNameWithoutExtension(typeof(TSettings));
+	public static string GetSettingsName<TSettings>(this ISettingsSink sink) where TSettings : ISettings
+		=> sink.GetSettingsName(typeof(TSettings));
 
 	/// <summary>
-	/// Gets the name for a settings class.
+	/// Gets the name of a settings class respecting the <see cref="SettingsNameAttribute"/>.
 	/// </summary>
 	/// <param name="sink"> The <see cref="ISettingsSink"/> that is extended. </param>
 	/// <param name="settingsType"> The type of the settings class. </param>
 	/// <returns> The name of the settings class. </returns>
-	public static string GetSettingsFileNameWithoutExtension(this ISettingsSink sink, Type settingsType)
+	public static string GetSettingsName(this ISettingsSink sink, Type settingsType)
 	{
 		if (!settingsType.IsClass) throw new ArgumentException($"The passed type '{settingsType}' must be a class.");
 		if (!typeof(ISettings).IsAssignableFrom(settingsType)) throw new ArgumentException($"The passed type '{settingsType}' must implement the interface '{nameof(ISettings)}'.");